add functionality to create and copy all folders, new lists that shows loaded folders, new buttons allowing changing the position of items in the list and deleting items, the structure of copied folders is retained.

+++ Views/CopyFilesView.xaml.cs
@@ -46,29 +46,79 @@ namespace CopyFilesByModificationDate.Views
 
         private async void CopyBtn_Click(object sender, RoutedEventArgs e)
         {
-            _destinationPath = DestinationPathTextBlock.Text;
-            bool result = await Task.Run( () => _copyFilesViewModel.CopyFiles(_destinationPath));
-            if (!result) ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Failed to copy files!"; else ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Files succesfully copied!";
+            if (SourceDirectoryListView.SelectedItems.Count != 0)
+            {
+                var dir = (DirectoryListItem)SourceDirectoryListView.SelectedItem;
+                _destinationPath = DestinationPathTextBlock.Text;
+                int result = await Task.Run(() => _copyFilesViewModel.CopyFilesAsync(_destinationPath, null, null));
+                if (result < 0) ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Failed to copy files!"; else ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Files succesfully copied!";
+            }
         }
 
         private void LoadItemsBtn_Click(object sender, RoutedEventArgs e)
         {
             _sourcePath = SourcePathTextBlock.Text;
-            bool result = _copyFilesViewModel.LoadItems(_sourcePath);
+            bool result = _copyFilesViewModel.LoadDirectories(_sourcePath);
             if (!result) ResultTextBlock.Text = "Failed to load files!"; else ResultTextBlock.Text = "Files succesfully loaded!";
         }
         
         private void DeleteSelectedItemBtn_Click(object sender, RoutedEventArgs e)
         {
+            if (FileListView.SelectedItems.Count != 0)
+            {
+                var item = FileListView.SelectedItem as FileListItem;
+                bool result;
+                if (item != null)
+                {
+                    result = _copyFilesViewModel.DeleteItem(item);
+                    if (result) ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Deleted item from list";
+                }
+                else ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Failed to delete item";
+            }
+        }
+
+        private void MoveUpBtn_Click(object sender, RoutedEventArgs e)
+        {
             var item = FileListView.SelectedItem as FileListItem;
             bool result;
             if (item != null)
             {
-                result = _copyFilesViewModel.DeleteItem(item);
-                if (result) ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Deleted item from list";
+                result = _copyFilesViewModel.MoveItemUp(item);
+                if (result) ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Moved item up";
             }
-            else ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Failed to delete item";
-            
+            else ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Failed to move item";
+            FileListView.SelectedItem = item;
+        }
+
+        private void MoveDownBtn_Click(object sender, RoutedEventArgs e)
+        {
+            var item = FileListView.SelectedItem as FileListItem;
+            bool result;
+            if (item != null)
+            {
+                result = _copyFilesViewModel.MoveItemDown(item);
+                if (result) ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Moved item down";
+            }
+            else ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Failed to move item";
+            FileListView.SelectedItem = item;
+        }
+
+        private async void CopyAllBtn_Click(object sender, RoutedEventArgs e)
+        {
+            _destinationPath = SourcePathTextBlock.Text;
+            _destinationPath = DestinationPathTextBlock.Text;
+            bool result = await Task.Run(() => _copyFilesViewModel.CopyAllFilesAsync(_sourcePath, _destinationPath));
+            if (!result) ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Failed to copy files!"; else ResultTextBlock.Text = ResultTextBlock.Text + "\n" + "Files succesfully copied!";
+        }
+
+        private async void SourceDirectoryListView_SelectedAsync(object sender, RoutedEventArgs e)
+        {
+            if (SourceDirectoryListView.SelectedItems.Count != 0) 
+            { 
+                var dir = (DirectoryListItem)SourceDirectoryListView.SelectedItem;
+                bool result = await _copyFilesViewModel.LoadItems(dir.FullPath);
+                if (!result) ResultTextBlock.Text = "\nFailed to load files from "+ dir.DirectoryName +" directory!"; else ResultTextBlock.Text = "\nFiles succesfully loaded from " + dir.DirectoryName + " directory!";
+            }
         }
     }
 }
